# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Preezma task-api",
    "description": "Artur Karapetyan",
    "license": {
      "name": "",
      "url": ""
    }
  },
  "host": "localhost:3001",
  "basePath": "/api",
  "tags": [
    {
      "name": "Clients",
      "description": "Clients API"
    },
    {
      "name": "Providers",
      "description": "Providers API"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/clients": {
      "post": {
        "tags": ["Clients"],
        "summary": "Create a new client",
        "parameters": [
          {
            "name": "client",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/createClient"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "New client has been created",
            "schema": {
              "$ref": "#/definitions/getClient"
            }
          },
          "409": {
            "description": "Duplication detected",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "5XX": {
            "description": "Unexpected error"
          }
        }
      },
      "get": {
        "tags": ["Clients"],
        "summary": "Get all clients",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/getClients"
            }
          },
          "5XX": {
            "description": "Unexpected error"
          }
        }
      }
    },
    "/client/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "The client ID",
          "type": "string"
        }
      ],
      "get": {
        "tags": ["Clients"],
        "summary": "Get client",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/getClient"
            }
          },
          "404": {
            "description": "Client not found",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "5XX": {
            "description": "Unexpected error"
          }
        }
      },
      "put": {
        "tags": ["Clients"],
        "summary": "Update client",
        "parameters": [
          {
            "name": "client",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/updateClient"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "The client has been updated",
            "schema": {
              "$ref": "#/definitions/getClient"
            }
          },
          "409": {
            "description": "Duplication detected",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "404": {
            "description": "Client not found",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "5XX": {
            "description": "Unexpected error"
          }
        }
      },
      "delete": {
        "summary": "Delete client",
        "tags": ["Clients"],
        "responses": {
          "200": {
            "description": "The client has been deleted",
            "schema": {
              "$ref": "#/definitions/deleteResponse"
            }
          },
          "404": {
            "description": "Client not found",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "5XX": {
            "description": "Unexpected error"
          }
        }
      }
    },
    "/provider": {
      "post": {
        "tags": ["Providers"],
        "summary": "Create a new provider",
        "parameters": [
          {
            "name": "provider",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/provider"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "New provider has been created",
            "schema": {
              "$ref": "#/definitions/getProvider"
            }
          },
          "409": {
            "description": "Duplication detected",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "5XX": {
            "description": "Unexpected error"
          }
        }
      }
    },

    "/providers": {
      "get": {
        "tags": ["Providers"],
        "summary": "Get all providers",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/getProviders"
            }
          },
          "5XX": {
            "description": "Unexpected error"
          }
        }
      }
    },
    "/provider/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "The provider ID",
          "type": "string"
        }
      ],
      "put": {
        "tags": ["Providers"],
        "summary": "Update provider",
        "parameters": [
          {
            "name": "provider",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/provider"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "The provider has been updated",
            "schema": {
              "$ref": "#/definitions/getProvider"
            }
          },
          "409": {
            "description": "Duplication detected",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "404": {
            "description": "Provider not found",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "5XX": {
            "description": "Unexpected error"
          }
        }
      },
      "delete": {
        "summary": "Delete provider",
        "tags": ["Providers"],
        "responses": {
          "200": {
            "description": "The provider has been deleted",
            "schema": {
              "$ref": "#/definitions/deleteResponse"
            }
          },
          "404": {
            "description": "Provider not found",
            "schema": {
              "$ref": "#/definitions/errorResponse"
            }
          },
          "5XX": {
            "description": "Unexpected error"
          }
        }
      }
    }
  },
  "definitions": {
    "createClient": {
      "required": ["name", "email", "phone"],
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "phone": {
          "type": "string"
        },
        "providers": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "updateClient": {
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "phone": {
          "type": "string"
        },
        "providers": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "getClient": {
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "phone": {
          "type": "string"
        },
        "providers": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "getClients": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/getClient"
      }
    },
    "provider": {
      "required": ["name"],
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "getProvider": {
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "getProviders": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/getProvider"
      }
    },
    "deleteResponse": {
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "errorResponse": {
      "properties": {
        "error": {
          "properties": {
            "message": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "success": {
              "type": "boolean"
            },
            "status": {
              "type": "number"
            }
          }
        }
      }
    }
  }
}